requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
click==7.1.2
Jinja2==3.0.0a1
MarkupSafe==2.0.0a1
monopacker==0.1.0
more-itertools==8.2.0
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
py==1.8.1
pyfakefs==4.0.2
pyparsing==2.4.7
pytest==5.4.1
regex==2020.5.7
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.9
